Remove early return for non commonjs environments
